Bug Fix: Mimic Clock
Mimic Clock by default runs a fake clock that starts at Epoch 0.
To get a real clock value so that Token Expirations will match
a real clock for external usage we need to set the '-r' option
when starting Mimic, thus allowing Repose to validate the tokens
based on the Expiration Time.