Add src subtree for pkg(s)
This is uncommon for python projects. I'm doing it to make packaging
easier to maintain and harder to break in non-obvious ways. The
difference is documented in the readme so as not to confuse future
developers.

There's a good article on how the conventional method can break here:
https://blog.ionelmc.ro/2014/05/25/python-packaging/#the-structure

See also comments on PR #48: https://github.com/rackerlabs/rse/pull/48