Less fragile cache init check
Our memcache library has a bug; if stats() returns values with
whitespaces, it thinks it's an invalid response. In most cases that
doesn't matter, but it seems our prod memcache reports the version as
"$VERSION $OS".

Instead we'll just get a nonexistent key. We don't care about the return
value, we just want it not to raise an exception.