Fix docker line-ending issue on windows hosts
When the repo is checked out on a windows machine and then docker is
run, the resulting files in docker have the wrong line endings. This
breaks certain parts of the dockerfile, in particular the ones that
install packages. Instead of installing a list of packages from a file,
they end up trying to install a single packages with embedded line feeds
in the name. That 1. breaks, and 2. produces *very* confusing error
messages.

This commit runs dos2unix on affected files before the dockerfile tries
to use them.