Allow full logging of HTTP traffic
This is implemented as a special logger, provided by util.py.

We are unlikely to need this often, and it eats disk space like mad.
Hence all http logging is done at the TRACE level and directed to a
separate file.

It would be better if I could wrap the controller, or the app object,
based on a configuration setting.  Unfortunately there is no single
location that has access to all of the request headers, response
headers, and body of each.

Note carefully the way the loglines are formed. RSE is
performance-sensitive. The `log.trace('%s', request)` usage keeps
requests from being stringified (which can be expensive) if they are not
going to be logged.