CBU-1438 - Implement ElasticAPM
RSE does not utilize an OOB supported WSGI framework for automatic
instrumentation. That said, it is required to manually begin and end
transactions. To start with, I have done this in the simplest way I
could find possible. We may want to try adding more spans to the
traces collected as well as additional context information to the
transactions, so that the agent can be more useful to us.

Unfortunately, a lot of information NewRelic was able to automatically
gather and report on regarding RSE transactions will not be available
at this time via Elastic APM.